Pre-processing eyetribe data
Process the data (based on the eyetracker_output.txt file) and get it ready to be used in ML algorithms. It takes the data and puts it into a pandas data frame.  It is important to note that ML aglorithms can only take numerical values. Thus, the columns 'Fix', 'State', and 'Timestamp' will not be used in ML but are included here for consistency. These columns don't seem that helpful for our assignment anyway.